orientar nuestro repositorio a terminos del dominio

+++ src/main/java/com/supermercado/domain/repository/ProductRepository.java
@@ -1,15 +1,16 @@
 package com.supermercado.domain.repository;
 
 import com.supermercado.domain.Product;
+import com.supermercado.persistence.entity.Producto;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface ProductRepository {
     List<Product> getAll();
-    Optional<List<Product>> getByCategory(int categoryId);
+    Optional<List<Product>> getByCategory(long categoryId);
     Optional<List<Product>> getScarseProduct(int quantity);
-    Optional<Product> getProduct(int productId);
+    Optional<Product> getProduct(long productId);
     Product save(Product product);
-    void delete(int productId);
+    void delete(long productId);
 }
